\name{R4DSXML-package}
\alias{R4DSXML-package}
\alias{R4DSXML}
\docType{package}
\title{
Read CDISC data files
}
\description{
R4DSXML package includs functions for reading Dataset-XML and Define-XML files.
}
\details{
\tabular{ll}{
Package: \tab R4DSXML\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-12-31\cr
License: \tab LGL-3\cr
}
R4DSXML package is provide functions to import Dataset-XML and Define-XML.
\cr
It is supported Dataset-XML ver. 1.0 and Define-XML ver. 2.0 in this package.
}
\source{
R4DSXML package includes example files both of Dataset-XML and Define-XML which were obtained from Dataset-XML 1.0 package on CDISC web site. \url{http://www.cdisc.org/dataset-xml}
}
\author{
Ippei Akiya\cr
Maintainer: Ippei Akiya <ippei.aki@datadriven.jp>
}
\references{
CDISC Dataset-XML Specification Version 1.0 (\url{http://www.cdisc.org/dataset-xml})\cr
CDISC Define-XML Specification Version 2.0 (\url{http://www.cdisc.org/define-xml})
}

\keyword{ CDISC }
